requirements.txt
fastapi
uvicorn
pillow
pytesseract
langchain
langchain-openai
python-dotenv
python-multipart
langchain-nvidia-ai-endpoints
langchain-mistralai